Ajustado miogrations

+++ src/Database/Migrations/2020_12_29_100230_update_emissora_table.php
@@ -8,19 +8,23 @@
 		protected $connection = 'sulradio';
 		
 		public function up() {
-			Schema::connection($this->connection)
-				->table('emissora', function( $collection) {
-					$collection->integer("client_id")->nullable()->index();
-					$collection->index("razao_social");
-					$collection->index("nome_fantasia");
-				});
+            if(Schema::connection($this->connection)->hasTable('emissora')) {
+                Schema::connection($this->connection)
+                    ->table('emissora', function ($collection) {
+                        $collection->integer("client_id")->nullable()->index();
+                        $collection->index("razao_social");
+                        $collection->index("nome_fantasia");
+                    });
+            }
 		}
 		
 		public function down() {
-			Schema::connection($this->connection)->table('emissora', function (Blueprint $table) {
-				$table->dropColumn(['client_id']);
-				$table->dropIndex(['razao_social']);
-				$table->dropIndex(['nome_fantasia']);
-			});
+            if(Schema::connection($this->connection)->hasTable('emissora')) {
+                Schema::connection($this->connection)->table('emissora', function (Blueprint $table) {
+                    $table->dropColumn(['client_id']);
+                    $table->dropIndex(['razao_social']);
+                    $table->dropIndex(['nome_fantasia']);
+                });
+            }
 		}
 	}
