+++ src/Database/Migrations/2021_02_23_193107_update_document_type_table.php
@@ -9,15 +9,19 @@ class UpdateDocumentTypeTable extends Migration {
 	protected $connection = 'sulradio';
 	
 	public function up() {
-		Schema::connection($this->connection)
-			->table('document_type', function(Blueprint $table) {
-				$table->enum('goal', ['Cliente', 'Engenharia', 'Jurídico', 'Administrativo'])->default('Cliente')->index();
-			});
+        if(Schema::connection($this->connection)->hasTable('document_type')) {
+            Schema::connection($this->connection)
+                ->table('document_type', function (Blueprint $table) {
+                    $table->enum('goal', ['Cliente', 'Engenharia', 'Jurídico', 'Administrativo'])->default('Cliente')->index();
+                });
+        }
 	}
 	
 	public function down() {
-		Schema::connection($this->connection)->table('document_type', function (Blueprint $table) {
-			$table->dropColumn('goal');
-		});
+        if(Schema::connection($this->connection)->hasTable('document_type')) {
+            Schema::connection($this->connection)->table('document_type', function (Blueprint $table) {
+                $table->dropColumn('goal');
+            });
+        }
 	}
 }
