+++ src/Database/Migrations/2021_02_23_194007_update_document_goal_table.php
@@ -9,17 +9,21 @@ class UpdateDocumentGoalTable extends Migration {
 	protected $connection = 'sulradio';
 	
 	public function up() {
-		Schema::connection($this->connection)
-			->table('document', function(Blueprint $table) {
-				$table->enum('goal', ['Cliente', 'Engenharia', 'Jurídico', 'Adminstrativo'])
-					->default('Cliente')
-					->index();
-			});
+        if(Schema::connection($this->connection)->hasTable('document')) {
+            Schema::connection($this->connection)
+                ->table('document', function (Blueprint $table) {
+                    $table->enum('goal', ['Cliente', 'Engenharia', 'Jurídico', 'Adminstrativo'])
+                        ->default('Cliente')
+                        ->index();
+                });
+        }
 	}
 	
 	public function down() {
-		Schema::connection($this->connection)->table('document', function (Blueprint $table) {
-			$table->dropColumn('goal');
-		});
+        if(Schema::connection($this->connection)->hasTable('document')) {
+            Schema::connection($this->connection)->table('document', function (Blueprint $table) {
+                $table->dropColumn('goal');
+            });
+        }
 	}
 }
