+++ src/Database/Migrations/2021_12_25_130330_update_emissora.php
@@ -8,18 +8,22 @@
 		protected $connection = 'sulradio';
 		
 		public function up() {
-			Schema::connection($this->connection)
-				->table('emissora', function( $collection) {
-					$collection->string("fistel", 50)->nullable()->index();
-					$collection->string("url_mosaico", 191)->nullable();
-					$collection->string("url_seacco", 191)->nullable();
-					$collection->string("url_cnpj", 191)->nullable();
-				});
+            if(Schema::connection($this->connection)->hasTable('emissora')) {
+                Schema::connection($this->connection)
+                    ->table('emissora', function ($collection) {
+                        $collection->string("fistel", 50)->nullable()->index();
+                        $collection->string("url_mosaico", 191)->nullable();
+                        $collection->string("url_seacco", 191)->nullable();
+                        $collection->string("url_cnpj", 191)->nullable();
+                    });
+            }
 		}
 		
 		public function down() {
-			Schema::connection($this->connection)->table('emissora', function (Blueprint $table) {
-				$table->dropColumn(['fistel', 'url_mosaico', 'url_seacco', 'url_cnpj']);
-			});
+            if(Schema::connection($this->connection)->hasTable('emissora')) {
+                Schema::connection($this->connection)->table('emissora', function (Blueprint $table) {
+                    $table->dropColumn(['fistel', 'url_mosaico', 'url_seacco', 'url_cnpj']);
+                });
+            }
 		}
 	}
